amount of product feature done per product
